+++ CHANGELOG.markdown
@@ -3,6 +3,7 @@ EDGE
 * Added like comments functionality
 * Added flow pagination with will_paginate
 * Graffities and replys through ajax
+* Collapse replys
 
 0.1.0
 ----